refactor: evict remaining Noir/Mavros coupling from common; close the axis
Relocate the last Noir-toolchain pieces out of the field-agnostic spine so
provekit-common no longer depends on acir/noirc_abi/mavros-vm/mavros-artifacts:

- NoirElement alias + noir_to_native -> provekit-field-bn254 (new noir_element
  module; field-bn254 gains an acir dep). The acir<->native bridge only
  type-checks when the native field is bn254, so it belongs in the bn254 crate
  and is naturally excluded from non-bn254 builds.
- PrintAbi -> provekit-r1cs-compiler (git-renamed; sole consumer).
- convert_mavros_r1cs_to_provekit -> provekit-r1cs-compiler (new mavros_convert
  module; uses mavros-artifacts, only consumer). InternedFieldElement stays
  private to common (inferred, not named).
- Delete the orphan utils/file_io.rs (never wired into the module tree).
- Drop acir/noirc_abi/mavros-vm/mavros-artifacts from common/Cargo.toml.

Bridge consumers (noir, r1cs-compiler hot path, wasm, bench) repointed to
provekit_field_bn254. common/src now references the Noir toolchain only in a
code comment. Determinism gate + field KATs stay green; verifier/verifier-server
pull provekit-noir as types-only (no nargo/witness-generation).
