+++ provekit/common/src/utils/mod.rs
@@ -1,12 +1,9 @@
-mod print_abi;
 pub mod serde_ark;
 pub mod serde_ark_option;
 pub mod serde_ark_vec;
 pub mod serde_hex;
 pub mod serde_jsonify;
 pub mod sumcheck;
-
-pub use self::print_abi::PrintAbi;
 
 /// Deserializes a BN254 field element from up to 32 little-endian bytes.
 #[inline]
@@ -28,7 +25,7 @@ pub fn field_to_bytes_le(fe: FieldElement) -> [u8; 32] {
 }
 
 use {
-    crate::{FieldElement, NoirElement},
+    crate::FieldElement,
     ark_ff::{BigInt, Field, PrimeField},
     ruint::{aliases::U256, uint},
     std::{
@@ -78,13 +75,6 @@ fn unzip_double_array<T: Sized, const N: usize, const M: usize>(
 
 pub const fn uint_to_field(i: U256) -> FieldElement {
     FieldElement::new(BigInt(i.into_limbs()))
-}
-
-/// Convert a Noir field element to a native `FieldElement`
-#[inline(always)]
-pub fn noir_to_native(n: NoirElement) -> FieldElement {
-    let limbs = n.into_repr().into_bigint().0;
-    FieldElement::from(BigInt(limbs))
 }
 
 /// Calculates the degree of the next smallest power of two
@@ -179,58 +169,4 @@ pub fn batch_inverse_montgomery(values: &[FieldElement]) -> Vec<FieldElement> {
     inverses[0] = inv_acc;
 
     inverses
-}
-
-#[cfg(not(target_arch = "wasm32"))]
-pub fn convert_mavros_r1cs_to_provekit(mavros_r1cs: &mavros_artifacts::R1CS) -> crate::R1CS {
-    let num_witnesses = mavros_r1cs.witness_layout.size();
-    let num_constraints = mavros_r1cs.constraints.len();
-
-    let total_entries: usize = mavros_r1cs
-        .constraints
-        .iter()
-        .map(|c| c.a.len() + c.b.len() + c.c.len())
-        .sum();
-
-    let mut r1cs = crate::R1CS::new();
-    r1cs.add_witnesses(num_witnesses);
-    r1cs.reserve_constraints(num_constraints, total_entries);
-
-    let mut a_buf: Vec<(u32, crate::InternedFieldElement)> = Vec::with_capacity(64);
-    let mut b_buf: Vec<(u32, crate::InternedFieldElement)> = Vec::with_capacity(64);
-    let mut c_buf: Vec<(u32, crate::InternedFieldElement)> = Vec::with_capacity(64);
-
-    for constraint in &mavros_r1cs.constraints {
-        a_buf.clear();
-        a_buf.extend(
-            constraint
-                .a
-                .iter()
-                .map(|(idx, coeff)| (*idx as u32, r1cs.intern(*coeff))),
-        );
-
-        b_buf.clear();
-        b_buf.extend(
-            constraint
-                .b
-                .iter()
-                .map(|(idx, coeff)| (*idx as u32, r1cs.intern(*coeff))),
-        );
-
-        c_buf.clear();
-        c_buf.extend(
-            constraint
-                .c
-                .iter()
-                .map(|(idx, coeff)| (*idx as u32, r1cs.intern(*coeff))),
-        );
-
-        r1cs.push_constraint(
-            a_buf.iter().copied(),
-            b_buf.iter().copied(),
-            c_buf.iter().copied(),
-        );
-    }
-
-    r1cs
 }
