+++ provekit/field/bn254/Cargo.toml
@@ -18,6 +18,7 @@ poseidon2.workspace = true
 skyscraper.workspace = true
 
 # Cryptography and proof systems
+acir.workspace = true
 ark-bn254.workspace = true
 ark-ff.workspace = true
 spongefish.workspace = true
