+++ provekit/r1cs-compiler/src/lib.rs
@@ -1,11 +1,13 @@
 mod binops;
 mod constraint_helpers;
 mod digits;
+mod mavros_convert;
 mod memory;
 pub mod msm;
 mod noir_proof_scheme;
 pub mod noir_to_r1cs;
 mod poseidon2;
+mod print_abi;
 pub mod range_check;
 mod sha256_compression;
 mod spread;
