+++ provekit/r1cs-compiler/src/noir_to_r1cs.rs
@@ -21,12 +21,12 @@ use {
     ark_ff::{BigInteger, PrimeField},
     ark_std::{One, Zero},
     provekit_common::{
-        utils::noir_to_native,
         witness::{
             ConstantOrR1CSWitness, ConstantTerm, SumTerm, WitnessBuilder, BINOP_ATOMIC_BITS,
         },
-        FieldElement, NoirElement, R1CS,
+        FieldElement, R1CS,
     },
+    provekit_field_bn254::{noir_to_native, NoirElement},
     std::{collections::BTreeMap, num::NonZeroU32, ops::Neg},
 };
